# lib/api-zod/src/generated/types/abilityListResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fusion Pokémon API for game bots, battle tools, and Pokédex clients.
 * OpenAPI spec version: 1.0.0
 */
import type { Ability } from './ability';

export interface AbilityListResponse {
  success: boolean;
  data: Ability[];
  count: number;
}