# lib/api-zod/src/generated/types/battleInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fusion Pokémon API for game bots, battle tools, and Pokédex clients.
 * OpenAPI spec version: 1.0.0
 */

export interface BattleInfo {
  role: string;
  tier: string;
  weight: number;
  catchRate: number;
}