# lib/api-zod/src/generated/types/battleResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fusion Pokémon API for game bots, battle tools, and Pokédex clients.
 * OpenAPI spec version: 1.0.0
 */
import type { BattleResult } from './battleResult';

export interface BattleResponse {
  success: boolean;
  data: BattleResult;
}