# lib/api-zod/src/generated/types/battleResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fusion Pokémon API for game bots, battle tools, and Pokédex clients.
 * OpenAPI spec version: 1.0.0
 */
import type { Fusion } from './fusion';
import type { Move } from './move';

export interface BattleResult {
  attacker: Fusion;
  defender: Fusion;
  recommendedMove: Move;
  matchup: string;
}