# lib/api-zod/src/generated/types/fusion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fusion Pokémon API for game bots, battle tools, and Pokédex clients.
 * OpenAPI spec version: 1.0.0
 */
import type { BattleInfo } from './battleInfo';
import type { FusionSprite } from './fusionSprite';
import type { StatBlock } from './statBlock';

export interface Fusion {
  id: string;
  name: string;
  parents: string[];
  types: string[];
  rarity: string;
  description: string;
  stats: StatBlock;
  abilities: string[];
  /** @nullable */
  hiddenAbility: string | null;
  weaknesses: string[];
  resistances: string[];
  immunities: string[];
  /**
     * @minItems 4
     * @maxItems 4
     */
  moves: string[];
  sprite: FusionSprite;
  image: string;
  battle: BattleInfo;
}