# lib/api-zod/src/generated/types/moveListResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fusion Pokémon API for game bots, battle tools, and Pokédex clients.
 * OpenAPI spec version: 1.0.0
 */
import type { Move } from './move';

export interface MoveListResponse {
  success: boolean;
  data: Move[];
  count: number;
  total?: number;
}